✨[feat]:  모델 에 값을 넘기게 수정
